more sound

+++ src/carts/files/main.ts
@@ -5,7 +5,7 @@ export function load(): void {
 }
 
 // called per-frame to update screen
-export function update(dt: i32): void {
+export function update(dt: u64): void {
 }
 
 // called when the cart is unloaded
